TOPAZ_v4.4
Add function to identify the silver flit and use silver tag in
arbitration. A flit is silver when it is circulated at least 1x in a
subring. However, no such flit is identified.  So, it doesn't help to
reduce the deflection rate.  Functions may be useful later.

Note: the actual injection rate = L*PktLength. So, the current
implementation start to save power when injection rate < 0.15.

However, dynamic power almost double when injection rate is 0.16.

This result is compared with CHIPPER.

Possible Solutions:
Use Surround XY. Need to forbid flit from being sent back. So PN and
routing need to be modified.