TOPAZ_v4.5
This implementation contains major improvements:

1. Not that m_portUtilCount and m_portState are inport indexed.
m_portType is outport indexed. So, the related index errors in
TPZSimpleRouterFlowBless::portPowerGate() and
TPZSimpleRouterFlowBless::inputReading() are corrected.

2. U-turn restriction is added in the second stage of permutation
network. A flit is allowed to make an U-turn only when the other
contending flits can
make productive progress. In TPZSimpleRouterFlowBless::permuterBlock(),
the swap signals in the second stage will be raised to disable U-turn.

3. Tried to use the following design in order to reduce deflection, but
it does NOT work.
In Section setRoutingPort, Y-dimension port is set as desired outPort
when the X-dimension candidate is not exhausted but disabled. If no
productive
port is available, simply assign a random desired port. Let the
permutation network handle the deflection. However, the overall power
increase about 8%.
The reason is UNKNOWN.