TOPAZ_v4.6
Modified Static Power Saving computation.

Assume vdd takes 15 cycles to fully discharge. (defined in Network.cpp)

If #_of_consecutive_sleep_cycle > 15, the saving on the begining of 15
cycles is calculated using the quadratic equation reported in
"Microarchitectural Techniques for Power Gating
of Execution Units". The rest of sleeping cycles will be substituted to
the linear equation.